chore: docker mount fix for CI and mac
