+++ test/bdd/fixtures/stakeholder-server/docker-compose.yml
@@ -21,16 +21,8 @@ services:
     ports:
       - "8088:8088"
     volumes:
-      - type: bind
-        source: ../wellknown/jws/did-trustbloc
-        target: /web/.well-known/did-trustbloc
-      - type: bind
-        source: ../wellknown/jws/stakeholder.one/did-configuration.json
-        target: /web/.well-known/did-configuration.json
-        volume:
-          nocopy: true
-        bind:
-          propagation: shared
+      - ../wellknown/jws/did-trustbloc:/web/.well-known/did-trustbloc
+      - ../wellknown/jws/stakeholder.one/did-configuration.json:/web/.well-known/did-configuration.json
       - ../keys/tls:/etc/tls
     networks:
       - did-method-rest_bdd_net
